+++ ts/object/src/lib/serializer.ts
@@ -103,8 +103,12 @@ function deflateObject( item:Object, params?:any ) {
         if ( m && m.properties.has(field) ) {
             /* ignore delegated properties */
             if ( m.property(field)['ʘdelegate'] ) continue
-            /* ignore inheritied properties */
-            if ( m.property(field)['ʘinherit'] ) continue
+            /* ignore unpopulated inheritied properties with undefined value */
+            if ( m.property(field)['ʘinherit']  ) {
+                if ( item[`ʘ${field}`] === undefined ) {
+                    continue
+                }
+           } 
 
             if (  m.property(field)['ʘcoerce'] ) {
                 const coerce = m.property(field)['ʘcoerce']